moving host checker
