+++ src/config.js
@@ -1,5 +1,5 @@
-const fs = require('fs')
-const path = require('path')
+import fs from 'fs'
+import path from 'path'
 
 export function initConfig () {
   return new Promise((resolve, reject) => {
