+++ src/parsePage.js
@@ -2,25 +2,6 @@ import { XMLHttpRequest } from 'xmlhttprequest'
 import { JSDOM } from 'jsdom'
 import { detailsParserBuilder } from './pageParserAdapter/index'
 
-export function checkWhichHost (URL) {
-  if (URL.startsWith('https://www.curseforge.com/wow/addons/')) {
-    return parseURLCurseforge(URL)
-  }
-  const errorObj = { 'error': "ERROR: Invalid URL '" + URL + "'. Given link does not match parse" }
-  return errorObj
-}
-
-// curseforge
-export function parseURLCurseforge (URL) {
-  const URLSplit = URL.split('https://www.curseforge.com/wow/addons/')
-  if (URLSplit.length === 2) {
-    const URLObj = { 'URL': URL, 'host': 'curseforge', 'name': URLSplit[1] }
-    return URLObj
-  }
-  const errorObj = { 'error': "ERROR: Invalid URL '" + URL + "'." }
-  return errorObj
-}
-
 export function parseAddonDetails (URLObj) {
   return new Promise((resolve, reject) => {
     let req = new XMLHttpRequest()
